packages/ingestion-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
httpx==0.25.2
requests==2.31.0
openai==1.3.0
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
pydub==0.25.1
pillow==10.1.0
pdf2image==1.16.3
pytesseract==0.3.10
